+++ src/components/SearchControls/index.js
@@ -1,0 +1,3 @@
+import SearchControls from './SearchControls';
+
+export default SearchControls;
