--- src/components/SearchControlsFilter/index.js
@@ -1,3 +1,0 @@
-import SearchControlsFilter from './SearchControlsFilter';
-
-export default SearchControlsFilter;
