--- src/components/SearchControlsSort/SearchControlsSort.stories.mdx
@@ -1,20 +1,0 @@
-import { Meta, Story, Canvas } from '@storybook/addon-docs';
-
-import SearchControlsSort from './SearchControlsSort';
-
-<Meta
-    title='Core/SearchControlsSort'
-    component={SearchControls}
-    args={{
-        type: 'dataset',
-        sortProps: { options: ['relevance', 'recentlyadded'], value: 'relevance' },
-        inputProps: {
-            value: 'covid 19',
-        },
-        width: '900px',
-    }}
-/>
-
-<Canvas>
-    <Story name='Basic'>{props => <SearchControlsSort {...props} />}</Story>
-</Canvas>
