--- src/components/SearchControlsSort/index.js
@@ -1,3 +1,0 @@
-import SearchControlsSort from './SearchControlsSort';
-
-export default SearchControlsSort;
