+++ src/pages/dashboard/Components/AccountDatasets/AccountDatasetsContent.js
@@ -3,7 +3,8 @@ import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import Icon from '../../../../components/Icon';
-import SearchControlsSort from '../../../../components/SearchControlsSort';
+import LayoutBox from '../../../../components/LayoutBox';
+import SearchControls from '../../../../components/SearchControls';
 import { DATASETS_STATUS_ACTIVE, STATUS_ARCHIVE, STATUS_INREVIEW, STATUS_REJECTED } from '../../../../configs/constants';
 import DatasetCard from '../../../commonComponents/DatasetCard';
 import MessageNotFound from '../../../commonComponents/MessageNotFound';
@@ -14,10 +15,18 @@ const options = ['latest', 'alphabetic', 'metadata'];
 const optionsWithPublish = ['latest', 'alphabetic', 'metadata', 'recentlyadded'];
 
 const AccountDatasetsContent = ({ data = [], onSubmit, onReset, isLoading, isFetched, status, params, team, count }) => {
+	const { search, sortBy, sortDirection, maxResults } = params;
+
 	const [searchValue, setSearchValue] = useState('');
+	const [sortValue, setSortValue] = useState({
+		value: sortBy,
+		direction: sortDirection,
+	});
+
 	const history = useHistory();
 
 	const { t } = useTranslation();
+
 	const handleChange = React.useCallback(value => {
 		setSearchValue(value);
 	}, []);
@@ -27,6 +36,32 @@ const AccountDatasetsContent = ({ data = [], onSubmit, onReset, isLoading, isFet
 	const handleActivityLogClick = id => {
 		history.push(getDatasetPath(id));
 	};
+
+	const handleKeyDownEnter = React.useCallback(submitForm => {
+		submitForm();
+	}, []);
+
+	const handleSort = React.useCallback(
+		(data, submitForm) => {
+			setSortValue(data);
+
+			submitForm();
+		},
+		[searchValue]
+	);
+
+	React.useEffect(() => {
+		setSearchValue(search);
+	}, [search]);
+
+	React.useEffect(() => {
+		if (sortBy && sortDirection) {
+			setSortValue({
+				value: sortBy,
+				direction: sortDirection,
+			});
+		}
+	}, [sortBy, sortDirection]);
 
 	const hasActivityHistory = React.useCallback(dataset => dataset.listOfVersions.length > 0 && team === 'admin', [team]);
 
@@ -47,8 +82,6 @@ const AccountDatasetsContent = ({ data = [], onSubmit, onReset, isLoading, isFet
 		return datasetCardProps;
 	};
 
-	const { search, sortBy, sortDirection, maxResults } = params;
-
 	let statusOptions = options;
 
 	if (team !== 'admin' && (status === STATUS_ARCHIVE || status === DATASETS_STATUS_ACTIVE)) {
@@ -58,20 +91,22 @@ const AccountDatasetsContent = ({ data = [], onSubmit, onReset, isLoading, isFet
 	return (
 		<>
 			{isFetched && (
-				<SearchControlsSort
+				<SearchControls
 					type={t(`dataset.${status}`)}
 					onSubmit={onSubmit}
 					inputProps={{
+						onKeyDownEnter: handleKeyDownEnter,
 						onChange: handleChange,
-						value: search,
+						value: searchValue,
 						onReset,
 					}}
 					sortProps={{
-						direction: sortDirection,
-						value: sortBy,
+						direction: sortValue.direction,
+						value: sortValue.value,
 						options: statusOptions,
 						allowDirection: true,
 						minWidth: '300px',
+						onSort: handleSort,
 					}}
 					mt={4}
 				/>
@@ -79,7 +114,11 @@ const AccountDatasetsContent = ({ data = [], onSubmit, onReset, isLoading, isFet
 
 			<SearchResults
 				data={data}
-				errorMessage={({ type }) => <MessageNotFound word={pluralize(type)} />}
+				errorMessage={({ type }) => (
+					<LayoutBox mt={2}>
+						<MessageNotFound word={pluralize(type)} />
+					</LayoutBox>
+				)}
 				results={data =>
 					data.map(dataset => (
 						<DatasetCard
